requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
SQLAlchemy==2.0.36
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
cryptography==43.0.3
paramiko==3.5.0
APScheduler==3.10.4
email-validator==2.2.0
httpx==0.27.2
jinja2==3.1.4
python-multipart==0.0.12
orjson==3.10.7
uvloop==0.20.0
redis==5.0.8
pyyaml==6.0.2
aiosmtplib==3.0.1
twilio==8.10.0
